pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stocks-langraph"
version = "0.1.0"
description = "Stateful LangGraph stock trading agent with approvals, checkpointing, and observability."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "langchain-openrouter>=0.1.0",
  "langgraph>=0.2.67",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "langsmith>=0.3.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
stocks-agent = "stocks_langgraph.cli:main"
stocks-api = "stocks_langgraph.api:run"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
